comments

+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/assignment/BalanceSubtopologyGraphConstructor.java
@@ -223,7 +223,7 @@ public class BalanceSubtopologyGraphConstructor implements RackAwareGraphConstru
                 }
             }
 
-            taskNodeId += taskIds.stream().filter(taskIdSet::contains).count();
+            taskNodeId += (int) taskIds.stream().filter(taskIdSet::contains).count();
             topicGroupIndex++;
         }
